# proto/goddess/goddess/api/v1/member_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: goddess/api/v1/member.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationMemberServiceDismissMember = "/goddess.api.v1.MemberService/DismissMember"
const OperationMemberServiceGetMember = "/goddess.api.v1.MemberService/GetMember"
const OperationMemberServiceInviteMember = "/goddess.api.v1.MemberService/InviteMember"
const OperationMemberServiceListMember = "/goddess.api.v1.MemberService/ListMember"
const OperationMemberServiceSelectMember = "/goddess.api.v1.MemberService/SelectMember"
const OperationMemberServiceUpdateMemberStatus = "/goddess.api.v1.MemberService/UpdateMemberStatus"

type MemberServiceHTTPServer interface {
	DismissMember(context.Context, *DismissMemberRequest) (*DismissMemberReply, error)
	GetMember(context.Context, *GetMemberRequest) (*MemberItem, error)
	InviteMember(context.Context, *InviteMemberRequest) (*InviteMemberReply, error)
	ListMember(context.Context, *ListMemberRequest) (*ListMemberReply, error)
	SelectMember(context.Context, *SelectMemberRequest) (*SelectMemberReply, error)
	UpdateMemberStatus(context.Context, *UpdateMemberStatusRequest) (*UpdateMemberStatusReply, error)
}

func RegisterMemberServiceHTTPServer(s *http.Server, srv MemberServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/member/invite", _MemberService_InviteMember0_HTTP_Handler(srv))
	r.DELETE("/v1/member/{uid}", _MemberService_DismissMember0_HTTP_Handler(srv))
	r.GET("/v1/members", _MemberService_ListMember0_HTTP_Handler(srv))
	r.GET("/v1/members/select", _MemberService_SelectMember0_HTTP_Handler(srv))
	r.GET("/v1/member/{uid}", _MemberService_GetMember0_HTTP_Handler(srv))
	r.PUT("/v1/member/{uid}/status", _MemberService_UpdateMemberStatus0_HTTP_Handler(srv))
}

func _MemberService_InviteMember0_HTTP_Handler(srv MemberServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in InviteMemberRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberServiceInviteMember)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.InviteMember(ctx, req.(*InviteMemberRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*InviteMemberReply)
		return ctx.Result(200, reply)
	}
}

func _MemberService_DismissMember0_HTTP_Handler(srv MemberServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DismissMemberRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberServiceDismissMember)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DismissMember(ctx, req.(*DismissMemberRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DismissMemberReply)
		return ctx.Result(200, reply)
	}
}

func _MemberService_ListMember0_HTTP_Handler(srv MemberServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListMemberRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberServiceListMember)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListMember(ctx, req.(*ListMemberRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListMemberReply)
		return ctx.Result(200, reply)
	}
}

func _MemberService_SelectMember0_HTTP_Handler(srv MemberServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SelectMemberRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberServiceSelectMember)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SelectMember(ctx, req.(*SelectMemberRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SelectMemberReply)
		return ctx.Result(200, reply)
	}
}

func _MemberService_GetMember0_HTTP_Handler(srv MemberServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMemberRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberServiceGetMember)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMember(ctx, req.(*GetMemberRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MemberItem)
		return ctx.Result(200, reply)
	}
}

func _MemberService_UpdateMemberStatus0_HTTP_Handler(srv MemberServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateMemberStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberServiceUpdateMemberStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateMemberStatus(ctx, req.(*UpdateMemberStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateMemberStatusReply)
		return ctx.Result(200, reply)
	}
}

type MemberServiceHTTPClient interface {
	DismissMember(ctx context.Context, req *DismissMemberRequest, opts ...http.CallOption) (rsp *DismissMemberReply, err error)
	GetMember(ctx context.Context, req *GetMemberRequest, opts ...http.CallOption) (rsp *MemberItem, err error)
	InviteMember(ctx context.Context, req *InviteMemberRequest, opts ...http.CallOption) (rsp *InviteMemberReply, err error)
	ListMember(ctx context.Context, req *ListMemberRequest, opts ...http.CallOption) (rsp *ListMemberReply, err error)
	SelectMember(ctx context.Context, req *SelectMemberRequest, opts ...http.CallOption) (rsp *SelectMemberReply, err error)
	UpdateMemberStatus(ctx context.Context, req *UpdateMemberStatusRequest, opts ...http.CallOption) (rsp *UpdateMemberStatusReply, err error)
}

type MemberServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewMemberServiceHTTPClient(client *http.Client) MemberServiceHTTPClient {
	return &MemberServiceHTTPClientImpl{client}
}

func (c *MemberServiceHTTPClientImpl) DismissMember(ctx context.Context, in *DismissMemberRequest, opts ...http.CallOption) (*DismissMemberReply, error) {
	var out DismissMemberReply
	pattern := "/v1/member/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMemberServiceDismissMember))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MemberServiceHTTPClientImpl) GetMember(ctx context.Context, in *GetMemberRequest, opts ...http.CallOption) (*MemberItem, error) {
	var out MemberItem
	pattern := "/v1/member/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMemberServiceGetMember))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MemberServiceHTTPClientImpl) InviteMember(ctx context.Context, in *InviteMemberRequest, opts ...http.CallOption) (*InviteMemberReply, error) {
	var out InviteMemberReply
	pattern := "/v1/member/invite"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMemberServiceInviteMember))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MemberServiceHTTPClientImpl) ListMember(ctx context.Context, in *ListMemberRequest, opts ...http.CallOption) (*ListMemberReply, error) {
	var out ListMemberReply
	pattern := "/v1/members"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMemberServiceListMember))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MemberServiceHTTPClientImpl) SelectMember(ctx context.Context, in *SelectMemberRequest, opts ...http.CallOption) (*SelectMemberReply, error) {
	var out SelectMemberReply
	pattern := "/v1/members/select"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMemberServiceSelectMember))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MemberServiceHTTPClientImpl) UpdateMemberStatus(ctx context.Context, in *UpdateMemberStatusRequest, opts ...http.CallOption) (*UpdateMemberStatusReply, error) {
	var out UpdateMemberStatusReply
	pattern := "/v1/member/{uid}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMemberServiceUpdateMemberStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}