# proto/goddess/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/member/invite:
        post:
            tags:
                - MemberService
            operationId: MemberService_InviteMember
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.InviteMemberRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.InviteMemberReply'
    /v1/member/{uid}:
        get:
            tags:
                - MemberService
            operationId: MemberService_GetMember
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.MemberItem'
        delete:
            tags:
                - MemberService
            operationId: MemberService_DismissMember
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.DismissMemberReply'
    /v1/member/{uid}/status:
        put:
            tags:
                - MemberService
            operationId: MemberService_UpdateMemberStatus
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.UpdateMemberStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.UpdateMemberStatusReply'
    /v1/members:
        get:
            tags:
                - MemberService
            operationId: MemberService_ListMember
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: userUID
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: memberUID
                  in: query
                  schema:
                    type: string
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: email
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.ListMemberReply'
    /v1/members/select:
        get:
            tags:
                - MemberService
            operationId: MemberService_SelectMember
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: lastUID
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.SelectMemberReply'
...
components:
    schemas:
...
        goddess.api.v1.DismissMemberReply:
            type: object
            properties:
                message:
                    type: string
...
        goddess.api.v1.InviteMemberReply:
            type: object
            properties:
                message:
                    type: string
        goddess.api.v1.InviteMemberRequest:
            type: object
            properties:
                email:
                    type: string
                roleUID:
                    type: integer
                    format: uint32
        goddess.api.v1.ListMemberReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/goddess.api.v1.MemberItem'
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
...
        goddess.api.v1.MemberItem:
            type: object
            properties:
                uid:
                    type: string
                email:
                    type: string
                status:
                    type: integer
                    format: enum
                createdAt:
                    type: string
                updatedAt:
                    type: string
                userUID:
                    type: string
                name:
                    type: string
                nickname:
                    type: string
                avatar:
                    type: string
                remark:
                    type: string
...
        goddess.api.v1.SelectMemberItem:
            type: object
            properties:
                value:
                    type: string
                label:
                    type: string
                disabled:
                    type: boolean
                tooltip:
                    type: string
        goddess.api.v1.SelectMemberReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/goddess.api.v1.SelectMemberItem'
                total:
                    type: string
                lastUID:
                    type: string
                hasMore:
                    type: boolean
...
        goddess.api.v1.UpdateMemberStatusReply:
            type: object
            properties:
                message:
                    type: string
        goddess.api.v1.UpdateMemberStatusRequest:
            type: object
            properties:
                uid:
                    type: string
                status:
                    type: integer
                    format: enum
...